pyproject.toml
[project]
name = "orbitfinalprojct"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.1.1",
    "crewai[tools]>=1.6.1",
    "python-dotenv>=1.2.2",
]
